requirements.txt
python-telegram-bot==11.1.0
redis==3.4
requests==2.23